fix: ai新窗口打开异常，结构性拆环
pageMaps → yakEditor → shortcutKeyCore（无回边），yakEditorPageShortcut → utils → pageMaps 独立，不再形成三角环。
